# docs/openapi.yaml
openapi: 3.0.1
info:
  title: 'EventXperience Backend API'
  description: 'Generated API reference for the EventXperience backend. Responses use the shared ApiResponse envelope unless noted otherwise.'
  version: v1
paths:
...
  '/api/events/clubs/{clubId}':
    get:
      tags:
        - events
      operationId: Events_GetEventsByClub
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            '$ref': '#/components/schemas/Int32'
        - name: status
          in: query
          schema:
            '$ref': '#/components/schemas/String7'
        - name: page
          in: query
          schema:
            '$ref': '#/components/schemas/Int325'
        - name: pageSize
          in: query
          schema:
            '$ref': '#/components/schemas/Int326'
      responses:
        200:
          description: OK
          content:
            text/plain:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfPagedResponseOfEventResponse'
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfPagedResponseOfEventResponse'
            text/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfPagedResponseOfEventResponse'
        400:
          description: 'Bad request, validation failure, or CSRF failure.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        500:
          description: 'Unexpected server error.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
...
  '/api/clubs/{clubId}/join':
    post:
      tags:
        - clubs
      operationId: Club_JoinClub
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            '$ref': '#/components/schemas/Int32'
      responses:
        200:
          description: OK
          content:
            text/plain:
              schema:
                '$ref': '#/components/schemas/MessageResponse'
            application/json:
              schema:
                '$ref': '#/components/schemas/MessageResponse'
            text/json:
              schema:
                '$ref': '#/components/schemas/MessageResponse'
        400:
          description: 'Bad request, validation failure, or CSRF failure.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        500:
          description: 'Unexpected server error.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        401:
          description: 'Authentication is required or the access token is invalid/expired.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        403:
          description: 'The authenticated user does not have permission to perform this action.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
      security:
        - bearerAuth:
            []
    delete:
      tags:
        - clubs
      operationId: Club_LeaveClub
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            '$ref': '#/components/schemas/Int32'
      responses:
        200:
          description: OK
          content:
            text/plain:
              schema:
                '$ref': '#/components/schemas/MessageResponse'
            application/json:
              schema:
                '$ref': '#/components/schemas/MessageResponse'
            text/json:
              schema:
                '$ref': '#/components/schemas/MessageResponse'
        400:
          description: 'Bad request, validation failure, or CSRF failure.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        500:
          description: 'Unexpected server error.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        401:
          description: 'Authentication is required or the access token is invalid/expired.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        403:
          description: 'The authenticated user does not have permission to perform this action.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
      security:
        - bearerAuth:
            []
  /api/clubs:
    post:
      tags:
        - clubs
      operationId: Club_CreateClub
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                Name:
                  '$ref': '#/components/schemas/String14'
                Description:
                  '$ref': '#/components/schemas/String14'
                Clubtype:
                  '$ref': '#/components/schemas/String15'
                ClubImage:
                  '$ref': '#/components/schemas/IFormFile'
                Phone:
                  '$ref': '#/components/schemas/String15'
                Email:
                  '$ref': '#/components/schemas/String15'
        required: true
      responses:
        201:
          description: Created
          content:
            text/plain:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfClubResponse'
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfClubResponse'
            text/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfClubResponse'
        400:
          description: 'Bad request, validation failure, or CSRF failure.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        500:
          description: 'Unexpected server error.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        401:
          description: 'Authentication is required or the access token is invalid/expired.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        403:
          description: 'The authenticated user does not have permission to perform this action.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
      security:
        - bearerAuth:
            []
    get:
      tags:
        - clubs
      operationId: Club_GetClubs
      parameters:
        - name: search
          in: query
          schema:
            '$ref': '#/components/schemas/String'
        - name: clubType
          in: query
          schema:
            '$ref': '#/components/schemas/String'
        - name: sortBy
          in: query
          schema:
            type: integer
            default: 0
        - name: page
          in: query
          schema:
            '$ref': '#/components/schemas/Int325'
        - name: pageSize
          in: query
          schema:
            '$ref': '#/components/schemas/Int326'
      responses:
        200:
          description: OK
          content:
            text/plain:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfPagedResponseOfClubResponse'
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfPagedResponseOfClubResponse'
            text/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfPagedResponseOfClubResponse'
        400:
          description: 'Bad request, validation failure, or CSRF failure.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        500:
          description: 'Unexpected server error.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
...
  /api/clubs/search:
    post:
      tags:
        - clubs
      operationId: Club_SearchClubs
      requestBody:
        content:
          application/json:
            schema:
              '$ref': '#/components/schemas/ClubSearchRequest'
          text/json:
            schema:
              '$ref': '#/components/schemas/ClubSearchRequest'
          'application/*+json':
            schema:
              '$ref': '#/components/schemas/ClubSearchRequest'
        required: true
      responses:
        200:
          description: OK
          content:
            text/plain:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfPagedResponseOfClubResponse'
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfPagedResponseOfClubResponse'
            text/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfPagedResponseOfClubResponse'
        400:
          description: 'Bad request, validation failure, or CSRF failure.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        500:
          description: 'Unexpected server error.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
  /api/clubs/managed:
    get:
      tags:
        - clubs
      operationId: Club_GetManagedClubs
      responses:
        200:
          description: OK
          content:
            text/plain:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfIEnumerableOfClubResponse'
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfIEnumerableOfClubResponse'
            text/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfIEnumerableOfClubResponse'
        400:
          description: 'Bad request, validation failure, or CSRF failure.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        500:
          description: 'Unexpected server error.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        401:
          description: 'Authentication is required or the access token is invalid/expired.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        403:
          description: 'The authenticated user does not have permission to perform this action.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
      security:
        - bearerAuth:
            []
  '/api/clubs/{id}/staff':
    get:
      tags:
        - clubs
      operationId: Club_GetClubStaff
      parameters:
        - name: id
          in: path
          required: true
          schema:
            '$ref': '#/components/schemas/Int32'
      responses:
        200:
          description: OK
          content:
            text/plain:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfIEnumerableOfClubStaffResponse'
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfIEnumerableOfClubStaffResponse'
            text/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfIEnumerableOfClubStaffResponse'
        400:
          description: 'Bad request, validation failure, or CSRF failure.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        500:
          description: 'Unexpected server error.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        401:
          description: 'Authentication is required or the access token is invalid/expired.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        403:
          description: 'The authenticated user does not have permission to perform this action.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
      security:
        - bearerAuth:
            []
...
components:
  schemas:
...
    ApiResponseOfClubResponse:
      type: object
      properties:
        success:
          '$ref': '#/components/schemas/Boolean'
        message:
          '$ref': '#/components/schemas/String'
        data:
          '$ref': '#/components/schemas/ClubResponse'
        error:
          '$ref': '#/components/schemas/ApiError'
        meta:
          '$ref': '#/components/schemas/Object'
...
    ApiResponseOfObject:
      type: object
      properties:
        success:
          '$ref': '#/components/schemas/Boolean'
        message:
          '$ref': '#/components/schemas/String'
        data:
          '$ref': '#/components/schemas/Object'
        error:
          '$ref': '#/components/schemas/ApiError'
        meta:
          '$ref': '#/components/schemas/Object'
...
    ApiResponseOfPagedResponseOfClubResponse:
      type: object
      properties:
        success:
          '$ref': '#/components/schemas/Boolean'
        message:
          '$ref': '#/components/schemas/String'
        data:
          '$ref': '#/components/schemas/PagedResponseOfClubResponse'
        error:
          '$ref': '#/components/schemas/ApiError'
        meta:
          '$ref': '#/components/schemas/Object'
...
    ApiResponseOfPagedResponseOfEventResponse:
      type: object
      properties:
        success:
          '$ref': '#/components/schemas/Boolean'
        message:
          '$ref': '#/components/schemas/String'
        data:
          '$ref': '#/components/schemas/PagedResponseOfEventResponse'
        error:
          '$ref': '#/components/schemas/ApiError'
        meta:
          '$ref': '#/components/schemas/Object'
...
    ClubSearchRequest:
      type: object
      properties:
        query:
          '$ref': '#/components/schemas/String2'
        filters:
          '$ref': '#/components/schemas/ClubSearchFilters'
        sortBy:
          '$ref': '#/components/schemas/ClubSortBy'
        page:
          '$ref': '#/components/schemas/Int329'
        pageSize:
          '$ref': '#/components/schemas/Int3210'
...
    IFormFile:
      type: string
      format: binary
    Int32:
      type: integer
      format: int32
...
    Int325:
      type: integer
      format: int32
      default: 1
    Int326:
      type: integer
      format: int32
      default: 20
...
    MessageResponse:
      type: object
      properties:
        success:
          '$ref': '#/components/schemas/Boolean'
        message:
          '$ref': '#/components/schemas/String'
        data:
          '$ref': '#/components/schemas/Object'
        error:
          '$ref': '#/components/schemas/ApiError'
        meta:
          '$ref': '#/components/schemas/Object'
...
    Object: {}
...
    String:
      type: string
...
    String14:
      maxLength: 30
      minLength: 0
      type: string
    String15:
      maxLength: 30
      minLength: 0
      pattern: '^(sport|sports|academic|social|cultural|game|gaming|music|other)$'
      type: string
...
    String7:
      type: string
      default: null
...
  securitySchemes:
    bearerAuth:
      type: http
      description: 'Send the access token in the Authorization header as `Bearer {token}` for protected API routes.'
      scheme: bearer
      bearerFormat: JWT
...